+++ workers/ohsome_quality_analyst/api/api.py
@@ -50,14 +50,12 @@ from ohsome_quality_analyst.api.response_models import (
 from ohsome_quality_analyst.config import configure_logging
 from ohsome_quality_analyst.definitions import (
     ATTRIBUTION_URL,
-    get_attribution,
     get_indicator_definitions,
     get_metadata,
     get_report_definitions,
     get_topic_definition,
     get_topic_definitions,
 )
-from ohsome_quality_analyst.geodatabase import client as db_client
 from ohsome_quality_analyst.projects.definitions import (
     ProjectEnum,
     get_project,
@@ -342,22 +340,6 @@ async def post_report(
     ).attribution()
     response.update(geojson_object)
     return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_GEOJSON)
-
-
-@app.get("/regions")
-async def get_available_regions(asGeoJSON: bool = False):
-    """Get regions as list of names and identifiers or as a GeoJSON."""
-    response = empty_api_response()
-    if asGeoJSON is True:
-        regions = await db_client.get_regions_as_geojson()
-        response.update(regions)
-        response["attribution"]["text"] = get_attribution(["OSM"])
-        return JSONResponse(
-            content=jsonable_encoder(response), media_type=MEDIA_TYPE_GEOJSON
-        )
-    else:
-        response["result"] = await db_client.get_regions()
-        return response
 
 
 @app.get(